+++ App.js
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import { NavigationContainer } from '@react-navigation/native';
 import {createStackNavigator} from '@react-navigation/stack';
-import { Start } from './Start';
-import { Game } from './Game';
+import { Start } from './pages/Start';
+import { Game } from './pages/Game';
 
 const Stack = createStackNavigator();
  
@@ -14,8 +14,7 @@ const App = () => {
         <Stack.Screen name='Game' component={Game} />
       </Stack.Navigator>
     </NavigationContainer>
-  )
-  
+  ) 
 }
 
 export default App
