+++ README.md
@@ -4,8 +4,7 @@ The aim fo this project was to learn using React Native and build my first ever 
 
 ## The problem
 
-
-
+For the project I decided to create an easy tilt based game. The objective of the game is to get the player to maneuver from point A to point B while avoiding obstacles by tilting the phone. To build the game I used react-native-game-engine, react-native-sensors and expo.io. For the styling I used styled components. If I had more time I would have created more levels (the game currently only has one level) and would also like the entities in the game to spawn in random places when starting the game. I would also have liekd to put some more styling into the game, to better match the styling with the story line of the game. Building the mobile game was a great learning experience and it's a project I would love to come back to and develop further!
 
 ## View it live
