--- pull_request_template.md
@@ -1,3 +1,0 @@
-My reflections on how this week's project turned out:
-
-Things I'd like to have clarified or explained in more detail:
